📝 chore: example code 수정
