--- src/features/auth/storage/index.ts
@@ -1,1 +1,0 @@
-export * from './authStorageData';
