+++ src/lib/axios.ts
@@ -1,0 +1,12 @@
+// example code
+
+// import {ASYNC_STORAGE_KEYS, asyncStorage} from '../store/local/asyncStorage';
+
+// axios.interceptors.request.use(async (config: AxiosRequestConfig) => {
+//   const token = await asyncStorage.get<string>(ASYNC_STORAGE_KEYS.AUTH_JWT);
+
+//   if (token) {
+//     config.headers.Authorization = `Bearer ${token}`;
+//   }
+//   return config;
+// });
